requirements.txt
fastapi
uvicorn
httpx
aiofiles
python-dotenv
python-multipart
websockets